Prebuster: fix delete_phpmyadmin script

+++ modules/prebuster/templates/delete_phpmyadmin.sh
@@ -4,6 +4,8 @@ echo "## Delete phpmyadmin #####################################################
 #delete phpmyadmin
 mysql --defaults-extra-file=/root/.my.cnf -e "drop database phpmyadmin;"
 mysql --defaults-extra-file=/root/.my.cnf -e "drop user phpmyadmin@localhost;"
-apt-get remove --purge phpmyadmin -y
+if apt-show-versions | grep phpmyadmin; then
+  apt-get remove --purge phpmyadmin -y
+fi
